Refactor update_repo_manager into galaxy.
