+++ lib/galaxy/app.py
@@ -23,6 +23,7 @@ from galaxy.model.database_heartbeat import DatabaseHeartbeat
 from galaxy.model.tags import GalaxyTagHandler
 from galaxy.queue_worker import GalaxyQueueWorker
 from galaxy.tool_shed.galaxy_install.installed_repository_manager import InstalledRepositoryManager
+from galaxy.tool_shed.galaxy_install.update_repository_manager import UpdateRepositoryManager
 from galaxy.tool_util.deps.views import DependencyResolversView
 from galaxy.tool_util.verify import test_data
 from galaxy.tools.cache import (
@@ -45,9 +46,6 @@ from galaxy.web import url_for
 from galaxy.web.proxy import ProxyManager
 from galaxy.web_stack import application_stack_instance
 from galaxy.webhooks import WebhooksRegistry
-from tool_shed.galaxy_install import (
-    update_repository_manager
-)
 
 log = logging.getLogger(__name__)
 app = None
@@ -138,7 +136,7 @@ class UniverseApplication(config.ConfiguresGalaxyMixin):
         # Load Data Manager
         self.data_managers = DataManagers(self)
         # Load the update repository manager.
-        self.update_repository_manager = update_repository_manager.UpdateRepositoryManager(self)
+        self.update_repository_manager = UpdateRepositoryManager(self)
         # Load proprietary datatype converters and display applications.
         self.installed_repository_manager.load_proprietary_converters_and_display_applications()
         # Load datatype display applications defined in local datatypes_conf.xml
