+++ lib/galaxy/tool_shed/galaxy_install/update_repository_manager.py
